requirements.txt
vocode
numpy
scipy
webrtcvad
python-dotenv
pydantic-settings
PyPDF2
speexdsp-python 